+++ src/vue/assets/Logo.vue
@@ -1,6 +1,6 @@
 <template>
   <router-link class="logotype" :to="vueRoutes.app">
-    <img :src="logoUrl" alt="TokenD">
+    <img :src="logoUrl" alt="INCLUSIVITY">
   </router-link>
 </template>
 
@@ -8,7 +8,7 @@
 import { vueRoutes } from '@/vue-router/routes'
 import { SchemeRegistry } from '@/modules-arch/scheme-registry'
 
-const DEFAULT_LOGO_URL = '/static/logo.svg'
+const DEFAULT_LOGO_URL = '/static/logo.png'
 
 export default {
   data: _ => ({