+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
+## Unreleased
+#### Fixed
+- Asset creation with zero initial pre-issued amount
+
 ## [1.10.2] - 2019-09-26
 
 ## [1.10.2-rc.0] - 2019-09-16