+++ CHANGELOG.md
@@ -9,6 +9,7 @@ for further information about branching and tagging conventions.
 ## [Unreleased]
 #### Fixed
 - A bug when the user can not enter a value in input-field with type 'number' on Firefox
+- A bug with infinite loading of the page "Register of shares"
 
 ## [1.10.3-rc.0] - 2019-10-11
 #### Added