+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
-## [Unreleased]
+## [1.10.3-rcinc.0] - 2019-10-17
 #### Fixed
 - A bug when the user can not enter a value in input-field with type 'number' on Firefox
 
@@ -1005,7 +1005,8 @@ for further information about branching and tagging conventions.
 
 ## [1.3.0] - 2019-03-01
 
-[Unreleased]: https://github.com/tokend/web-client/compare/1.10.3-rc.0...HEAD
+[Unreleased]: https://github.com/tokend/web-client/compare/1.10.3-rcinc.0...HEAD
+[1.10.3-rcinc.0]: https://github.com/tokend/web-client/compare/1.10.3-rc.0...1.10.3-rcinc.0
 [1.10.3-rc.0]: https://github.com/tokend/web-client/compare/1.10.2...1.10.3-rc.0
 [1.10.2]: https://github.com/tokend/web-client/compare/1.10.2-rc.0...1.10.2
 [1.10.2-rc.0]: https://github.com/tokend/web-client/compare/1.10.1...1.10.2-rc.0